+++ meshroom/core/mtyping.py
